fixed the select nutrient bug
